+++ js/admin.js
@@ -2,8 +2,7 @@ import {
   campoRequerido,
   validarFormularioJuego,
   validarURL,
-  crearFila,
-} from "./functionsLoginGames.js";
+} from "./validations.js";
 
 import { Game } from "./gameClass.js";
 
@@ -95,6 +94,31 @@ function crearJuego() {
   juegoExistente = false;
   guardarJuegosEnLocalStorage();
   crearFila(juegoNuevo);
+}
+
+function crearFila(juegoNuevo) {
+  let tablaJuegos = document.getElementById("tablaJuegos");
+  let fila = document.createElement("tr");
+  let celda1 = document.createElement("th");
+  let celda2 = document.createElement("th");
+  let celda3 = document.createElement("th");
+  let celda4 = document.createElement("th");
+  let celda5 = document.createElement("th");
+  let celda6 = document.createElement("th");
+  celda1.innerHTML = `${juegoNuevo.codigo}`;
+  celda2.innerHTML = `${juegoNuevo.nombre}`;
+  celda3.innerHTML = `${juegoNuevo.categoria}`;
+  celda4.innerHTML = `${juegoNuevo.descripcion}`;
+  celda5.innerHTML = `${juegoNuevo.publicado}`;
+  celda6.innerHTML = ``;
+  fila.appendChild(celda1);
+  fila.appendChild(celda2);
+  fila.appendChild(celda3);
+  fila.appendChild(celda4);
+  fila.appendChild(celda5);
+  fila.appendChild(celda6);
+
+  tablaJuegos.appendChild(fila);
 }
 
 function modificarJuego() {}
