+++ js/validations.js
@@ -32,6 +32,17 @@ const validarRepetirContrasena = (input) => {
   }
 };
 
+const validarURL = (input) => {
+  let patron = /^https?:\/\/[\w\-]+(\.[\w\-]+)+[/#?]?.*$/;
+  if (patron.test(input.value)) {
+    input.className = "form-control is-valid";
+    return true;
+  } else {
+    input.className = "form-control is-invalid";
+    return false;
+  }
+};
+
 const validacionGeneral = (
   campoNombre,
   campoApellido,
@@ -55,9 +66,49 @@ const validacionGeneral = (
   }
 };
 
+const validarFormularioJuego = (nombre, categoria, descripcion, url) => {
+  if (
+    campoRequerido(nombre) &&
+    campoRequerido(categoria) &&
+    campoRequerido(descripcion) &&
+    validarURL(url)
+  ) {
+    return true;
+  } else {
+    return false;
+  }
+};
+
+const generalValidation = (
+  campoNombreApellido,
+  campoAsunto,
+  campoEmail,
+  campoObservaciones
+) => {
+  let alert = document.querySelector("#mensajeAlert");
+  Swal.fire(
+    "Bien Hecho!",
+    "Tu consulta fue enviada al administrador!",
+    "success"
+  );
+  if (
+    campoRequerido(campoNombreApellido) &&
+    campoRequerido(campoAsunto) &&
+    campoRequerido(campoObservaciones) &&
+    validarEmail(campoEmail)
+  ) {
+  } else {
+    alert.className = "alert alert-danger my-3";
+    return false;
+  }
+};
+
 export {
   campoRequerido,
   validarEmail,
   validarRepetirContrasena,
   validacionGeneral,
+  validarURL,
+  validarFormularioJuego,
+  generalValidation,
 };
