+++ newGame.html
@@ -172,7 +172,9 @@
                 <span class="circle" aria-hidden="true">
                   <span class="icon arrow"></span>
                 </span>
-                <span class="button-text text-white" id="ItemLogin">Regístrate!</span>
+                <span class="button-text text-white" id="ItemLogin"
+                  >Regístrate!</span
+                >
               </a>
             </button>
             <div class="btn-container">
@@ -305,7 +307,7 @@
         </div>
       </section>
     </main>
-    <script src="js/loginGame.js" type="module"></script>
+    <script src="js/admin.js" type="module"></script>
     <script src="js/bootstrap.min.js"></script>
     <script
       src="https://kit.fontawesome.com/7d2e342abf.js"
